Add support for versioning of Node and Edge classes
