+++ ReasoningEngine/Core/Edge.cs
@@ -1,19 +1,50 @@
-// Class to represent an edge in the graph
 namespace ReasoningEngine
 {
-    public class Edge
+    public class EdgeV1 : EdgeBase
     {
-        public long FromNode { get; set; }
-        public long ToNode { get; set; }
+        public override int Version => 1;
         public double Weight { get; set; }
-        public string EdgeContent { get; set; } = string.Empty;
+        public string EdgeContent { get; set; }
 
-        public Edge(long fromNode, long toNode, double weight, string edgeContent)
+        public EdgeV1(long fromNode, long toNode, double weight, string edgeContent) 
+            : base(fromNode, toNode)
         {
-            FromNode = fromNode;
-            ToNode = toNode;
             Weight = weight;
             EdgeContent = edgeContent;
         }
+
+        public override EdgeBase UpgradeToLatest()
+        {
+            return new EdgeV2(this.FromNode, this.ToNode, this.Weight, this.EdgeContent);
+        }
+    }
+
+    public class EdgeV2 : EdgeBase
+    {
+        public override int Version => 2;
+        public double Weight { get; set; }
+        public string EdgeContent { get; set; }
+        // Placeholder for future additions
+        // public DateTime CreationDate { get; set; }
+
+        public EdgeV2(long fromNode, long toNode, double weight, string edgeContent) 
+            : base(fromNode, toNode)
+        {
+            Weight = weight;
+            EdgeContent = edgeContent;
+            // CreationDate = DateTime.Now;
+        }
+
+        public override EdgeBase UpgradeToLatest()
+        {
+            return this; // Already the latest version
+        }
+    }
+
+    // Alias the latest version as Edge for easier use
+    public class Edge : EdgeV2
+    {
+        public Edge(long fromNode, long toNode, double weight, string edgeContent) 
+            : base(fromNode, toNode, weight, edgeContent) { }
     }
 }
