+++ ReasoningEngine/Core/IVersioned.cs
@@ -1,0 +1,7 @@
+namespace ReasoningEngine
+{
+    public interface IVersioned
+    {
+        int Version { get; }
+    }
+}
