+++ ReasoningEngine/GraphAlgorithms/GraphAlgorithmsStub.cs
@@ -1,0 +1,21 @@
+using System;
+using System.Collections.Generic;
+using System.Reflection;
+
+namespace ReasoningEngine.GraphAlgorithms
+{
+    public class GraphAlgorithmsStub
+    {
+        [VersionCompatibility(minimumNodeVersion: 2, minimumEdgeVersion: 1)]
+        public void AdvancedAlgorithm(NodeBase node, EdgeBase edge)
+        {
+            if (!VersionCompatibilityChecker.IsCompatible(node, MethodBase.GetCurrentMethod() as MethodInfo) ||
+                !VersionCompatibilityChecker.IsCompatible(edge, MethodBase.GetCurrentMethod() as MethodInfo))
+            {
+                throw new InvalidOperationException("Node or Edge version not compatible with this algorithm");
+            }
+
+            // Algorithm implementation
+        }
+    }
+}
